Slightly modified the the function trailCrossings to get rid of the `SelfVectorSpace b' constraints everywhere.
The signFromDerivAt (sub)function did calculations with vectors, and `abused'
tuples as V2 vectors, requiring (b,b) to be a vector space and finally b to be
a vector space. When using real vectors we only need that (V2 b) is a vector
space, and for that we need that b is an additive group.